backend/pyproject.toml
[tool.poetry]
name = "CECA-venv"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.72.0"
uvicorn = "^0.17.0"
SQLAlchemy = "^1.4.29"
mysql-connector-python = "^8.0.27"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pycodestyle = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
